Fix pause ambience loop not playing at fail screen
Matches stable.

---

Addresses https://github.com/ppy/osu/discussions/37580.
